Move uploader to robots
