+++ robots/cmd/uploader/BUILD.bazel
@@ -3,10 +3,10 @@ load("@io_bazel_rules_go//go:def.bzl", "go_binary", "go_library")
 go_library(
     name = "go_default_library",
     srcs = ["uploader.go"],
-    importpath = "kubevirt.io/project-infra/plugins/cmd/uploader",
+    importpath = "kubevirt.io/project-infra/robots/cmd/uploader",
     visibility = ["//visibility:private"],
     deps = [
-        "//plugins/mirror:go_default_library",
+        "//robots/pkg/mirror:go_default_library",
         "@com_github_bazelbuild_buildtools//build:go_default_library",
         "@com_google_cloud_go_storage//:go_default_library",
         "@org_golang_google_api//option:go_default_library",
