+++ robots/pkg/mirror/BUILD.bazel
@@ -3,7 +3,7 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 go_library(
     name = "go_default_library",
     srcs = ["bazel.go"],
-    importpath = "kubevirt.io/project-infra/plugins/mirror",
+    importpath = "kubevirt.io/project-infra/robots/pkg/mirror",
     visibility = ["//visibility:public"],
     deps = [
         "@com_github_bazelbuild_buildtools//build:go_default_library",
